# api.yml
openapi: 3.0.0
info:
  title: Integration API - G2A.COM
  version: 3.367.0
  description: |
    API for selling products on g2a.com
servers:
  - url: https://sandboxapi.g2a.com
    description: Sandbox server for testing
  - url: https://api.g2a.com
    description: Production server
...
paths:
...
  /v4/sales/orders:
    get:
      tags:
        - Seller Orders
      summary: Get seller orders list
      description: Returns seller orders list
      operationId: getSellerOrdersV4
      parameters:
        - $ref: 'parameters.yml#/OrderId'
        - $ref: 'parameters.yml#/CreatedAtFrom'
        - $ref: 'parameters.yml#/CreatedAtTo'
        - $ref: 'parameters.yml#/CustomerId'
        - $ref: 'parameters.yml#/Statuses'
        - $ref: 'parameters.yml#/ProductName'
        - $ref: 'parameters.yml#/SortBy'
        - $ref: 'parameters.yml#/SortStrategy'
        - $ref: 'parameters.yml#/Page'
        - $ref: 'parameters.yml#/ItemsPerPage'
      responses:
        '200':
          $ref: 'responses.yml#/GetSellerOrdersV4200'
        '400':
          $ref: 'responses.yml#/BadRequestV2'
        '500':
          $ref: 'responses.yml#/InternalServerErrorV2'
        '502':
          $ref: 'responses.yml#/BadGatewayV2'
  /v4/sales/orders/{orderId}:
    parameters:
      - $ref: 'parameters.yml#/OrderIdInPath'
    get:
      tags:
        - Seller Orders
      summary: Get seller order
      description: Returns seller order details
      operationId: getSellerOrderV4
      responses:
        '200':
          $ref: 'responses.yml#/GetSellerOrderV4200'
        '404':
          $ref: 'responses.yml#/NotFoundErrorV2'
        '500':
          $ref: 'responses.yml#/InternalServerErrorV2'
        '502':
          $ref: 'responses.yml#/BadGatewayV2'
...
components:
  securitySchemes:
    oAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /oauth/token
          scopes: { }
security:
  - oAuth2: [ ]